Move Configuration class into its own package
